create first controller api call to get all locations and working run app dev
