+++ src/server/src/controllers/location.ts
@@ -1,0 +1,34 @@
+/** source/controllers/location.ts */
+import { Request, Response, NextFunction } from 'express';
+import axios, { AxiosResponse } from 'axios';
+import { getLocations } from '../services/locationService'
+
+interface Post {
+    userId: Number;
+    id: Number;
+    title: String;
+    body: String;
+}
+
+// getting all posts
+const getPosts = async (req: Request, res: Response, next: NextFunction) => {
+    // get some posts
+    const locations = await getLocations()
+    return res.status(200).json({
+        locations
+    });
+};
+
+// getting a single post
+const getPost = async (req: Request, res: Response, next: NextFunction) => {
+    // get the post id from the req
+    let id: string = req.params.id;
+    // get the post
+    let result: AxiosResponse = await axios.get(`https://jsonplaceholder.typicode.com/posts/${id}`);
+    let post: Post = result.data;
+    return res.status(200).json({
+        message: post
+    });
+};
+
+export default { getPosts, getPost };
