+++ src/server/src/db/connect.ts
@@ -1,20 +1,27 @@
 import { Sequelize } from 'sequelize';
 import { credentials } from '../../../../dbCredentials.js';
 
-const sequelize = new Sequelize(credentials.database, credentials.user, credentials.password, {
+export const sequelize = new Sequelize(credentials.database, credentials.user, credentials.password, {
     host: credentials.host,
-    dialect: 'postgres'
-});
+    dialect: 'postgres',
+    define: {
+        freezeTableName: true
+      },
+    dialectOptions: {
+        ssl: {
+            require: true,
+            rejectUnauthorized: false 
+        }
+    }
+  });
 
-function connectToDB() {
+export function connectToDB(sequelize: Sequelize) {
     try {
         sequelize.authenticate();
         console.log('Connection has been established successfully.');
-        } catch (error) {
-            console.error('Unable to connect to the database:', error);
-        }
-
+    } catch (error) {
+        console.error('Unable to connect to the database:', error);
         sequelize.close();
+    }
 }
 
-connectToDB()
