+++ src/server/src/services/locationService.ts
@@ -1,0 +1,6 @@
+import { location } from '../db/models/location'
+
+export async function getLocations() {
+    console.log("Getting locations")
+    return await location.findAll();
+}
